+++ src/alltdl.common/Cmd/CmdRun.cs
@@ -1,6 +1,0 @@
-﻿namespace alltdl.Cmd
-{
-    public class CmdRun
-    {
-    }
-}